Restructure OTA module.
