+++ setup.py
@@ -13,6 +13,7 @@ setup(
     license="GPL-3.0",
     packages=find_packages(exclude=['*.tests']),
     install_requires=[
+        'aiohttp',
         'pycryptodome',
         'crccheck',
     ],
